+++ include/nucleus_generator.hpp
@@ -19,10 +19,10 @@ public:
     struct nucleus_params 
     {
          //Pb by default
-        const uint NA{208};
-        const uint ZA{82};
-        const uint NB{208};
-        const uint ZB{82};
+        const uint_fast16_t NA{208};
+        const uint_fast16_t ZA{82};
+        const uint_fast16_t NB{208};
+        const uint_fast16_t ZB{82};
          //Overall params
         const spatial min_distance{0.4};
         const bool shift_cms{true};
@@ -33,7 +33,7 @@ public:
 protected:
     static coords throw_nucleon_coords(std::shared_ptr<std::mt19937> random_generator, std::shared_ptr<ars> radial_sampler) noexcept;
     static double throw_radial(const double & random_number) noexcept;
-    static void throw_neutrons(std::vector<nucleon> *const nucleus, const uint & Z, std::shared_ptr<std::mt19937> random_generator) noexcept;
+    static void throw_neutrons(std::vector<nucleon> *const nucleus, const uint_fast16_t & Z, std::shared_ptr<std::mt19937> random_generator) noexcept;
     static bool coords_fit(const coords& co, const std::vector<coords>& other_coords, const spatial& min_distance) noexcept;
 private:
     static std::uniform_real_distribution<double> unif_dist;
